describing what motivates this separate repo
while I'm not sure we'll stick with it, this could stand as a repository for scripts that help create the backend, but don't necessarily get packaged with the application. A subset of scripts here could ship w/ the app, though, for example database creation scripts..... It's tough to say what's most sensible until we develop more.